requirements.txt
allure_python_commons==2.9.45
fastapi==0.82.0
loguru==0.4.1
psycopg2==2.9.3
pytest==6.2.5
PyYAML==6.0
requests==2.25.1
xlrd==2.0.1
xlwt==1.3.0
